Particles and RUNE (#12)
* Added GPU-based particle simulation and rendering through the Particle Emitter component
* Introduction of RUNE (Really Useful Node Editor) as a blueprint-like framework.
* Added initial foundation of Animation Blueprints (for animation state graph management)
* Started working on the Animation Window for interacting with animation blueprints.
* Modified premake files to make adding dependencies super easy
* Added imgui_node_editor library for RUNE
* Added magic_enum library to stringify enums for the UI
* Removed the FBX sdk - will introduced an FBX -> GLTF pipeline later
* Fixed a memory leak when resizing the scene or game view windows
* Fixed a memory corruption crash related to RawBuffers
* Refactored and cleaned up vulkan buffers to be more consolidated

+++ Projects/Engine/Source/Graphics/Rendering/Cubemap.cpp
@@ -62,9 +62,9 @@ namespace Odyssey
 		combinedBuffer.AppendData(backBuffer.GetData());
 
 		m_TextureDescription.ImageType = ImageType::Cubemap;
-		m_TextureDescription.Width = resolution;
-		m_TextureDescription.Height = resolution;
-		m_TextureDescription.Channels = channels;
+		m_TextureDescription.Width = (uint32_t)resolution;
+		m_TextureDescription.Height = (uint32_t)resolution;
+		m_TextureDescription.Channels = (uint32_t)channels;
 		m_TextureDescription.ArrayDepth = 6;
 
 		m_Texture = ResourceManager::Allocate<VulkanTexture>(m_TextureDescription, combinedBuffer);
